revise index.js handleNoteSave to add console logging for testing, add save button event listener
